Update data types and serialization benchmarks
The data types in the SBE schema have been updated from float to double and from uint16 to uint32 for better precision. We've also introduced new serializations benchmarks for the Jackson, Protobuf, and SBE serializers and deserializers, as well as updated the logging framework from Log4J to SLF4j for simplicity.