Reduce ByteBuffer allocation and improve serialization
The size of the ByteBuffer allocation in the SbeRecordSupplier class has been reduced from 1024 to 26. The serialization process in the SbeSerializer class has also been streamlined. It now directly returns the byte array when the byteBuffer has an array and its length equals the stockTradeEncoder's limit, thereby reducing the memory footprint and improving performance.